Add documentation for StArMap Client
This commit implements the automatic documentation generation for
StArMap client whenever the `main` branch is updated.

It uses Sphinx to generate the static HTML pages and GitLab CI to
deploy the pages to GitLab Pages.

It also adjust some docstrings to have a common layout as well as to
better suit the documentation generated by Sphinx.