Add alternative provider for StArMap client
This commit introduces the concept of `providers` which works as a way
to store and retrieve `QueryResponse` objects from a local environment
instead of having to request it to the server.

It also implements the first provider named `InMemoryMapProvider` which
aims to provide an "offline" way to store existing `QueryResponse`
objects in memory and use them to provide mappings. Then, whenever an
object exists "offline" (in memory) it will be returned instead of
having to request it from StArMap server.

Refers to SPSTRAT-325