Allow injecting custom session into StarmapClient
This commit changes the `StarmapClient` to allow receiving a custom
session which complies with the interface `StarmapBaseSession`.

It also changes the `StarmapClient` signature to have `url` not required
when the `session` is given by injection.

This change is useful to use the `StarmapClient` in offline mode, it is,
combined with a local provider in a scenario which network connection is
not possible.

Refers to SPSTRAT-326