Use the GH actions to update dependencies
This commit adds a new action named `pip-compile` which will daily
attempt to update the dependencies and open a new MR when necessary.

The advantage of doing this is that we can easily group the dependencies
without having to rely on complex iterations with dependabot.

This approach is the very same used in neighbor projects, such as
pubtools-ami:

https://github.com/release-engineering/pubtools-ami/blob/main/.github/workflows/pip-compile.yml

Refers to SPSTRAT-355