Implement support for APIv2 on Client
This commit changes the `StarmapClient` to be able to handle the APIv2
calls to retrieve a `QueryResponseContainer` whenever `v2` endpoints are
invoked.

It also changes the default `api_version` to `v2` and update the tests.

At this moment, the client implementation allows intercepting both APIv1
and APIv2 responses with the exception of the offline provider which
still only supports APIv1.

Refers to SPSTRAT-382